+++ src/components/button/Button.style.ts
@@ -1,0 +1,16 @@
+import { makeStyles } from 'tss-react/mui';
+
+const styles = {
+  root: {
+    backgroundColor: 'red',
+  }
+};
+
+export default makeStyles()(styles);
+
+/*
+* Inline Documentation
+*
+* examples and usage:
+* https://www.npmjs.com/package/tss-react
+* */
