+++ src/components/button/Button.tsx
@@ -1,23 +1,24 @@
-import {Button as MuiButton} from "@mui/material";
-import {ReactNode} from "react";
-import useStyles from './button.style.ts';
+import { Button as MuiButton } from '@mui/material';
+import { ReactNode } from 'react';
+import useStyles from './Button.style.ts';
+
 interface Props {
-    variant?: 'contained';
-    onClick?: () => void;
-    children?: ReactNode;
+  variant?: 'contained';
+  onClick?: () => void;
+  children?: ReactNode;
 }
 
 const Button = (props: Props) => {
-    const { classes } = useStyles();
-    const {variant, onClick, children } = props;
-    return (
-        <MuiButton
-            className={classes.root}
-            onClick={onClick}
-            variant={variant}>
-                {children}
-        </MuiButton>
-    );
+  const { classes } = useStyles();
+  const { variant, onClick, children } = props;
+  return (
+    <MuiButton
+      className={classes.root}
+      onClick={onClick}
+      variant={variant}>
+      {children}
+    </MuiButton>
+  );
 };
 
 export default Button;
