--- src/components/button/button.style.ts
@@ -1,16 +1,0 @@
-import { makeStyles} from "tss-react/mui";
-
-const styles = {
-    root: {
-        backgroundColor: 'red',
-    }
-}
-
-export default makeStyles()(styles)
-
-/*
-* Inline Documentation
-*
-* examples and usage:
-* https://www.npmjs.com/package/tss-react
-* */
